requirements.txt
certifi==2020.12.5
docx2txt==0.8
Pillow==8.0.1
